refactor: 특정상점조회 테스트코드 수정
