+++ src/test/java/in/koreatech/koin/acceptance/OwnerShopApiTest.java
@@ -43,9 +43,6 @@ import io.restassured.response.Response;
 class OwnerShopApiTest extends AcceptanceTest {
 
     @Autowired
-    private MenuCategoryRepository menuCategoryRepository;
-
-    @Autowired
     private OwnerRepository ownerRepository;
 
     @Autowired
@@ -249,90 +246,6 @@ class OwnerShopApiTest extends AcceptanceTest {
                 softly.assertThat(imageUrls).containsAnyElementsOf(shopImages.stream()
                     .map(ShopImage::getImageUrl).toList());
                 softly.assertThat(shopOpens).hasSize(2);
-            }
-        );
-    }
-
-    @Test
-    @DisplayName("특정 상점 조회")
-    void getShop() {
-        // given
-
-        ShopOpen open1 = ShopOpen.builder()
-            .openTime(LocalTime.of(9, 0))
-            .closeTime(LocalTime.of(21, 0))
-            .shop(shop)
-            .closed(false)
-            .dayOfWeek("MONDAY")
-            .build();
-
-        ShopOpen open2 = ShopOpen.builder()
-            .openTime(LocalTime.of(10, 0))
-            .closeTime(LocalTime.of(20, 30))
-            .shop(shop)
-            .closed(false)
-            .dayOfWeek("FRIDAY")
-            .build();
-        List<ShopOpen> shopOpens = shopOpenRepository.saveAll(List.of(open1, open2));
-
-        ShopCategoryMap shopCategoryMap1 = ShopCategoryMap.builder()
-            .shop(shop)
-            .shopCategory(shopCategory1)
-            .build();
-
-        ShopCategoryMap shopCategoryMap2 = ShopCategoryMap.builder()
-            .shop(shop)
-            .shopCategory(shopCategory2)
-            .build();
-
-        List<ShopCategoryMap> shopCategoryMaps = shopCategoryMapRepository.saveAll(
-            List.of(shopCategoryMap1, shopCategoryMap2));
-
-        ShopImage shopImage1 = ShopImage.builder()
-            .imageUrl("https://test.com/test1.jpg")
-            .shop(shop)
-            .build();
-
-        ShopImage shopImage2 = ShopImage.builder()
-            .imageUrl("https://test.com/test2.jpg")
-            .shop(shop)
-            .build();
-        shopImageRepository.saveAll(List.of(shopImage1, shopImage2));
-
-        ExtractableResponse<Response> response = RestAssured
-            .given()
-            .header("Authorization", "Bearer " + token)
-            .when()
-            .get("/shops/1")
-            .then()
-            .log().all()
-            .statusCode(HttpStatus.OK.value())
-            .extract();
-
-        List<ShopImage> savedShopImages = shopImageRepository.findAllByShopId(shop.getId());
-        List<MenuCategory> savedMenuCategories = menuCategoryRepository.findAllByShopId(shop.getId());
-        List<ShopOpen> savedShopOpens = shopOpenRepository.findAllByShopId(shop.getId());
-        List<ShopCategoryMap> savedShopCategoryMaps = shopCategoryMapRepository.findAllByShopId(shop.getId());
-
-        assertSoftly(
-            softly -> {
-                softly.assertThat(response.body().jsonPath().getString("address")).isEqualTo(shop.getAddress());
-                softly.assertThat(response.body().jsonPath().getBoolean("delivery")).isEqualTo(shop.getDelivery());
-                softly.assertThat(response.body().jsonPath().getLong("delivery_price"))
-                    .isEqualTo(shop.getDeliveryPrice());
-                softly.assertThat(response.body().jsonPath().getString("description")).isEqualTo(shop.getDescription());
-                softly.assertThat(response.body().jsonPath().getLong("id")).isEqualTo(shop.getId());
-                softly.assertThat(response.body().jsonPath().getString("name")).isEqualTo(shop.getName());
-                softly.assertThat(response.body().jsonPath().getBoolean("pay_bank")).isEqualTo(shop.getPayBank());
-                softly.assertThat(response.body().jsonPath().getBoolean("pay_card")).isEqualTo(shop.getPayCard());
-                softly.assertThat(response.body().jsonPath().getString("phone")).isEqualTo(shop.getPhone());
-
-                softly.assertThat(response.body().jsonPath().getList("image_urls")).hasSize(savedShopImages.size());
-                softly.assertThat(response.body().jsonPath().getList("menu_categories"))
-                    .hasSize(savedMenuCategories.size());
-                softly.assertThat(response.body().jsonPath().getList("open")).hasSize(savedShopOpens.size());
-                softly.assertThat(response.body().jsonPath().getList("shop_categories"))
-                    .hasSize(savedShopCategoryMaps.size());
             }
         );
     }
